Patch for AE compatibility
Condensed all scripts into cachemod, removed tampermod, added cache/cacheBase cross-compatible code, changed to internal.path.resolve edits for better mod compatibility, added "required": [] for AE compatibility.
